02_Data_Ingestion/scripts/requirements.txt
pandas==2.1.4
requests==2.31.0
kaggle==1.6.6
PyYAML==6.0
